+++ src/database/cipsrv_nhdplus_m/cipsrv_nhdplus_m_deploy.sql
@@ -7442,8 +7442,8 @@ BEGIN
           a.nhdplusid
          ,a.permanent_identifier 
          ,a.reachcode
-         ,ROUND(a.frommeas,5) AS fmeasure
-         ,ROUND(a.tomeas,5)   AS tmeasure
+         ,ROUND(a.frommeas::NUMERIC,5) AS fmeasure
+         ,ROUND(a.tomeas::NUMERIC,5)   AS tmeasure
          ,a.hydroseq
          ,a.uphydroseq
          ,a.terminalpa
@@ -7476,8 +7476,8 @@ BEGIN
           a.nhdplusid
          ,a.permanent_identifier 
          ,a.reachcode
-         ,ROUND(a.frommeas,5) AS fmeasure
-         ,ROUND(a.tomeas,5)   AS tmeasure
+         ,ROUND(a.frommeas::NUMERIC,5) AS fmeasure
+         ,ROUND(a.tomeas::NUMERIC,5)   AS tmeasure
          ,a.hydroseq
          ,a.uphydroseq
          ,a.terminalpa
@@ -7536,8 +7536,8 @@ BEGIN
           a.nhdplusid
          ,a.permanent_identifier 
          ,a.reachcode
-         ,ROUND(a.frommeas,5) AS fmeasure
-         ,ROUND(a.tomeas,5)   AS tmeasure
+         ,ROUND(a.frommeas::NUMERIC,5) AS fmeasure
+         ,ROUND(a.tomeas::NUMERIC,5)   AS tmeasure
          ,a.hydroseq
          ,a.uphydroseq
          ,a.terminalpa
@@ -7570,8 +7570,8 @@ BEGIN
           a.nhdplusid
          ,a.permanent_identifier 
          ,a.reachcode
-         ,ROUND(a.frommeas,5) AS fmeasure
-         ,ROUND(a.tomeas,5)   AS tmeasure
+         ,ROUND(a.frommeas::NUMERIC,5) AS fmeasure
+         ,ROUND(a.tomeas::NUMERIC,5)   AS tmeasure
          ,a.hydroseq
          ,a.uphydroseq
          ,a.terminalpa
@@ -7630,8 +7630,8 @@ BEGIN
           a.nhdplusid
          ,a.permanent_identifier 
          ,a.reachcode
-         ,ROUND(a.frommeas,5) AS fmeasure
-         ,ROUND(a.tomeas,5)   AS tmeasure
+         ,ROUND(a.frommeas::NUMERIC,5) AS fmeasure
+         ,ROUND(a.tomeas::NUMERIC,5)   AS tmeasure
          ,a.hydroseq
          ,a.uphydroseq
          ,a.terminalpa
@@ -7664,8 +7664,8 @@ BEGIN
           a.nhdplusid
          ,a.permanent_identifier 
          ,a.reachcode
-         ,ROUND(a.frommeas,5) AS fmeasure
-         ,ROUND(a.tomeas,5)   AS tmeasure
+         ,ROUND(a.frommeas::NUMERIC,5) AS fmeasure
+         ,ROUND(a.tomeas::NUMERIC,5)   AS tmeasure
          ,a.hydroseq
          ,a.uphydroseq
          ,a.terminalpa
@@ -7728,8 +7728,8 @@ BEGIN
           a.nhdplusid
          ,a.permanent_identifier 
          ,a.reachcode
-         ,ROUND(a.frommeas,5) AS fmeasure
-         ,ROUND(a.tomeas,5)   AS tmeasure
+         ,ROUND(a.frommeas::NUMERIC,5) AS fmeasure
+         ,ROUND(a.tomeas::NUMERIC,5)   AS tmeasure
          ,a.hydroseq
          ,a.uphydroseq
          ,a.terminalpa
@@ -7769,8 +7769,8 @@ BEGIN
           a.nhdplusid
          ,a.permanent_identifier 
          ,a.reachcode
-         ,ROUND(a.frommeas,5) AS fmeasure
-         ,ROUND(a.tomeas,5)   AS tmeasure
+         ,ROUND(a.frommeas::NUMERIC,5) AS fmeasure
+         ,ROUND(a.tomeas::NUMERIC,5)   AS tmeasure
          ,a.hydroseq
          ,a.uphydroseq
          ,a.terminalpa
