+++ src/database/cipsrv_nhdplus_m/functions/get_measure.sql
@@ -82,8 +82,8 @@ BEGIN
           a.nhdplusid
          ,a.permanent_identifier 
          ,a.reachcode
-         ,ROUND(a.frommeas,5) AS fmeasure
-         ,ROUND(a.tomeas,5)   AS tmeasure
+         ,ROUND(a.frommeas::NUMERIC,5) AS fmeasure
+         ,ROUND(a.tomeas::NUMERIC,5)   AS tmeasure
          ,a.hydroseq
          ,a.uphydroseq
          ,a.terminalpa
@@ -116,8 +116,8 @@ BEGIN
           a.nhdplusid
          ,a.permanent_identifier 
          ,a.reachcode
-         ,ROUND(a.frommeas,5) AS fmeasure
-         ,ROUND(a.tomeas,5)   AS tmeasure
+         ,ROUND(a.frommeas::NUMERIC,5) AS fmeasure
+         ,ROUND(a.tomeas::NUMERIC,5)   AS tmeasure
          ,a.hydroseq
          ,a.uphydroseq
          ,a.terminalpa
@@ -176,8 +176,8 @@ BEGIN
           a.nhdplusid
          ,a.permanent_identifier 
          ,a.reachcode
-         ,ROUND(a.frommeas,5) AS fmeasure
-         ,ROUND(a.tomeas,5)   AS tmeasure
+         ,ROUND(a.frommeas::NUMERIC,5) AS fmeasure
+         ,ROUND(a.tomeas::NUMERIC,5)   AS tmeasure
          ,a.hydroseq
          ,a.uphydroseq
          ,a.terminalpa
@@ -210,8 +210,8 @@ BEGIN
           a.nhdplusid
          ,a.permanent_identifier 
          ,a.reachcode
-         ,ROUND(a.frommeas,5) AS fmeasure
-         ,ROUND(a.tomeas,5)   AS tmeasure
+         ,ROUND(a.frommeas::NUMERIC,5) AS fmeasure
+         ,ROUND(a.tomeas::NUMERIC,5)   AS tmeasure
          ,a.hydroseq
          ,a.uphydroseq
          ,a.terminalpa
@@ -270,8 +270,8 @@ BEGIN
           a.nhdplusid
          ,a.permanent_identifier 
          ,a.reachcode
-         ,ROUND(a.frommeas,5) AS fmeasure
-         ,ROUND(a.tomeas,5)   AS tmeasure
+         ,ROUND(a.frommeas::NUMERIC,5) AS fmeasure
+         ,ROUND(a.tomeas::NUMERIC,5)   AS tmeasure
          ,a.hydroseq
          ,a.uphydroseq
          ,a.terminalpa
@@ -304,8 +304,8 @@ BEGIN
           a.nhdplusid
          ,a.permanent_identifier 
          ,a.reachcode
-         ,ROUND(a.frommeas,5) AS fmeasure
-         ,ROUND(a.tomeas,5)   AS tmeasure
+         ,ROUND(a.frommeas::NUMERIC,5) AS fmeasure
+         ,ROUND(a.tomeas::NUMERIC,5)   AS tmeasure
          ,a.hydroseq
          ,a.uphydroseq
          ,a.terminalpa
@@ -368,8 +368,8 @@ BEGIN
           a.nhdplusid
          ,a.permanent_identifier 
          ,a.reachcode
-         ,ROUND(a.frommeas,5) AS fmeasure
-         ,ROUND(a.tomeas,5)   AS tmeasure
+         ,ROUND(a.frommeas::NUMERIC,5) AS fmeasure
+         ,ROUND(a.tomeas::NUMERIC,5)   AS tmeasure
          ,a.hydroseq
          ,a.uphydroseq
          ,a.terminalpa
@@ -409,8 +409,8 @@ BEGIN
           a.nhdplusid
          ,a.permanent_identifier 
          ,a.reachcode
-         ,ROUND(a.frommeas,5) AS fmeasure
-         ,ROUND(a.tomeas,5)   AS tmeasure
+         ,ROUND(a.frommeas::NUMERIC,5) AS fmeasure
+         ,ROUND(a.tomeas::NUMERIC,5)   AS tmeasure
          ,a.hydroseq
          ,a.uphydroseq
          ,a.terminalpa
